Refactored quran_book_settings part 01
