+++ app/lib/modules/quran_book/quran_book_settings/cubit/quran_book_settings_cubit.dart
@@ -1,17 +1,22 @@
+import 'dart:convert';
+
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:mq_app_ui/mq_app_ui.dart';
+import 'package:mq_storage/mq_storage.dart';
+import 'package:my_quran/modules/modules.dart';
 
 part 'quran_book_settings_state.dart';
 
 class QuranBookSettingsCubit extends Cubit<QuranBookSettingsState> {
-  QuranBookSettingsCubit(this.repository) : super(const QuranBookSettingsState());
+  QuranBookSettingsCubit(this.storage) : super(const QuranBookSettingsState());
 
-  final ReadThemeRepository repository;
+  final PreferencesStorage storage;
+
+  static const _readThemeKey = 'readThemeKey';
 
   void init() {
-    final data = repository.getInitialThemeState();
+    final data = getInitialThemeState();
     emit(
       QuranBookSettingsState(
         modeIndex: data.modeIndex == 0 ? 0 : 1,
@@ -51,6 +56,21 @@ class QuranBookSettingsCubit extends Cubit<QuranBookSettingsState> {
       textSize: state.textSize,
       fontTypeIndex: state.fontType.indexValue,
     );
-    await repository.saveThemeState(date);
+    await saveThemeState(date);
+  }
+
+  ReadThemeData getInitialThemeState() {
+    try {
+      final value = storage.readString(key: _readThemeKey);
+      if (value == null) return ReadThemeData.initial;
+      return ReadThemeData.fromJson(json.decode(value) as Map<String, dynamic>);
+    } on Object catch (_) {
+      return ReadThemeData.initial;
+    }
+  }
+
+  Future<void> saveThemeState(ReadThemeData themeState) async {
+    final value = json.encode(themeState.toJson());
+    await storage.writeString(key: _readThemeKey, value: value);
   }
 }
