+++ app/lib/modules/quran_book/quran_book_settings/cubit/quran_book_settings_state.dart
@@ -1,43 +1,5 @@
 part of 'quran_book_settings_cubit.dart';
 
-enum QuranFontType {
-  complex,
-  tajweed,
-  uthmanic;
-
-  factory QuranFontType.fromIndex(int index) {
-    return switch (index) {
-      0 => QuranFontType.complex,
-      1 => QuranFontType.tajweed,
-      2 => QuranFontType.uthmanic,
-      _ => QuranFontType.complex,
-    };
-  }
-
-  int get indexValue {
-    return switch (this) {
-      QuranFontType.complex => 0,
-      QuranFontType.tajweed => 1,
-      QuranFontType.uthmanic => 2,
-    };
-  }
-}
-
-const _frReadThemeColor = [
-  Color(0xff000000),
-  Color(0xff000000),
-];
-
-const bgReadThemeColor = [
-  Color(0xffFFFFFF),
-  Color(0xffFFEED6),
-];
-
-const _loadingThemeColor = [
-  Color.fromARGB(255, 213, 210, 210),
-  Color.fromARGB(255, 241, 216, 180),
-];
-
 final class QuranBookSettingsState extends Equatable {
   const QuranBookSettingsState({
     this.verticalSpaceSize = 0,
@@ -80,25 +42,25 @@ final class QuranBookSettingsState extends Equatable {
 
   Color get frColor {
     try {
-      return _frReadThemeColor[modeIndex];
+      return ReadThemeData.frReadThemeColor[modeIndex];
     } on Object catch (_) {
-      return _frReadThemeColor[1];
+      return ReadThemeData.frReadThemeColor[1];
     }
   }
 
   Color get bgColor {
     try {
-      return bgReadThemeColor[modeIndex];
+      return ReadThemeData.bgReadThemeColor[modeIndex];
     } on Object catch (_) {
-      return bgReadThemeColor[1];
+      return ReadThemeData.bgReadThemeColor[1];
     }
   }
 
   Color get loadingColor {
     try {
-      return _loadingThemeColor[modeIndex];
+      return ReadThemeData.loadingThemeColor[modeIndex];
     } on Object catch (_) {
-      return _loadingThemeColor[1];
+      return ReadThemeData.loadingThemeColor[1];
     }
   }
 }
