+++ app/lib/modules/quran_book/quran_book_settings/models/quran_book_theme_data.dart
@@ -1,10 +1,4 @@
-import 'package:flutter/widgets.dart';
-
-abstract class ReadThemeRepository {
-  ReadThemeData getInitialThemeState();
-
-  Future<void> saveThemeState(ReadThemeData themeState);
-}
+import 'package:flutter/material.dart';
 
 @immutable
 final class ReadThemeData {
@@ -26,6 +20,30 @@ final class ReadThemeData {
     );
   }
 
+  final int modeIndex;
+  final double verticalSpaceSize;
+  final double horizontalSpaceSize;
+  final double textSize;
+  final int fontTypeIndex;
+
+  static const frReadThemeColor = [
+    Color(0xff000000),
+    Color(0xff000000),
+  ];
+
+  static const bgReadThemeColor = [
+    Color(0xffFFFFFF),
+    Color(0xffFFEED6),
+  ];
+
+  static const loadingThemeColor = [
+    Color.fromARGB(255, 213, 210, 210),
+    Color.fromARGB(255, 241, 216, 180),
+  ];
+
+  Color get foregroundColor => frReadThemeColor[modeIndex];
+  Color get backgroundColor => bgReadThemeColor[modeIndex];
+
   static const initial = ReadThemeData(
     modeIndex: 1,
     verticalSpaceSize: 0,
@@ -33,12 +51,6 @@ final class ReadThemeData {
     textSize: 28,
     fontTypeIndex: 0,
   );
-
-  final int modeIndex;
-  final double verticalSpaceSize;
-  final double horizontalSpaceSize;
-  final double textSize;
-  final int fontTypeIndex;
 
   Map<String, dynamic> toJson() {
     return <String, dynamic>{
@@ -49,22 +61,4 @@ final class ReadThemeData {
       'fontTypeIndex': fontTypeIndex,
     };
   }
-
-  static const frReadThemeColor = [
-    Color(0xff000000),
-    Color(0xff000000),
-    Color(0xffFFFFFF),
-    Color(0xffFFFFFF),
-  ];
-
-  Color get foregroundColor => frReadThemeColor[modeIndex];
-
-  static const bgReadThemeColor = [
-    Color(0xffFFFFFF),
-    Color(0xffFFEED6),
-    Color(0xff1C1C1E),
-    Color(0xff0E0E0F),
-  ];
-
-  Color get backgroundColor => bgReadThemeColor[modeIndex];
 }
