+++ app/lib/modules/quran_book/quran_book_settings/quran_book_settings.dart
@@ -1,3 +1,5 @@
 export 'cubit/quran_book_settings_cubit.dart';
+export 'models/quran_book_theme_data.dart';
+export 'models/quran_font_type.dart';
 export 'view/quran_book_setting_sheet.dart';
 export 'widgets/quran_fonts_selection_widget.dart';
