--- packages/mq_app_ui/lib/src/repo/sources/read_theme_local_data_source.dart
@@ -1,7 +1,0 @@
-import 'package:mq_app_ui/mq_app_ui.dart';
-
-abstract class LocalThemeDataSource {
-  ReadThemeData getInitialThemeState();
-
-  Future<void> saveThemeState(ReadThemeData themeState);
-}
